requirements.txt
backtrader>=1.9.78
pandas>=2.0.0
numpy>=1.24.0
smartapi-python>=1.3.0
pyotp>=2.9.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
